+++ controllers/productController.js
@@ -1,6 +1,7 @@
-const fs = require('fs');
+// const fs = require('fs');
+const Product = require('../models/productModel');
 
-let products = JSON.parse(
+/*let products = JSON.parse(
   fs.readFileSync(`${__dirname}/../dev-data/products.json`),
 );
 
@@ -19,7 +20,7 @@ exports.checkID = (req, res, next, val) => {
   req.product = product;
 
   next();
-};
+};*/
 
 exports.checkBody = (req, res, next) => {
   const { name, price } = req.body;
@@ -35,10 +36,10 @@ exports.checkBody = (req, res, next) => {
 
 exports.getAllProducts = (req, res) => {
   res.status(200).json({
-    results: products.length,
+    // results: products.length,
     requestedAt: req.requestTime,
     data: {
-      products: products,
+      // products: products,
     },
   });
 };
@@ -46,78 +47,31 @@ exports.getAllProducts = (req, res) => {
 exports.getProduct = (req, res) => {
   res.status(200).json({
     data: {
-      product: req.product,
+      // product: req.product,
     },
   });
 };
 
 exports.createProduct = (req, res) => {
-  const newId = products[products.length - 1].id + 1;
-  const newProduct = {
-    id: newId,
-    ...req.body,
-  };
-
-  products = [...products, newProduct];
-
-  fs.writeFile(
-    `${__dirname}/dev-data/products.json`,
-    JSON.stringify(products),
-    (err) => {
-      res.status(201).json({
-        data: {
-          product: { newProduct },
-        },
-      });
+  res.status(201).json({
+    data: {
+      // product: { newProduct },
     },
-  );
+  });
 };
 
 exports.updateProduct = (req, res) => {
-  const { product } = req;
-  const updatedProduct = {
-    ...product,
-    ...req.body,
-  };
-
-  const updatedProducts = products.map((prod) =>
-    prod.id == updatedProduct.id ? updatedProduct : prod,
-  );
-
-  fs.writeFile(
-    `${__dirname}/dev-data/products.json`,
-    JSON.stringify(updatedProducts),
-    (err) => {
-      res.status(200).json({
-        data: {
-          product: updatedProduct,
-        },
-      });
+  res.status(200).json({
+    data: {
+      // product: updatedProduct,
     },
-  );
+  });
 };
 
 exports.deleteProduct = (req, res) => {
-  const { id } = req.params;
-  const product = products.find((item) => item.id == id);
-
-  if (!product) {
-    return res.status(404).json({
-      msg: `${id} invalid id`,
-    });
-  }
-
-  const updatedProducts = products.filter((prod) => prod.id !== id);
-
-  fs.writeFile(
-    `${__dirname}/dev-data/products.json`,
-    JSON.stringify(updatedProducts),
-    (err) => {
-      res.status(200).json({
-        data: {
-          product: { product },
-        },
-      });
+  res.status(200).json({
+    data: {
+      // product: { product },
     },
-  );
+  });
 };
